abstracted out functions from WordBox in anticipation that i might need to use them later for the logical part of the typing test
