+++ src/WordBox/index.jsx
@@ -1,68 +1,19 @@
 import React from 'react';
 import '../index.css';
-import { Text, Flex } from '@chakra-ui/core';
-import words from './words.json';
+import { Flex } from '@chakra-ui/core';
 
-const WordBox = () => {
-  const getWords = (numWords) => {
-    const numW = numWords;
-    let typingWords = [];
+const WordBox = (props) => {
+  const words = Array.from(props.words);
 
-    typingWords = Array.from(
-      // make an array of the length passed in from random
-      // indeces of the words file
-      { length: numW },
-      () => words.english[Math.floor(Math.random() * words.english.length)]
-    );
-
-    console.log(typingWords);
-
-    return typingWords;
-  };
-
-  const highlightCurrWord = (current) => {
-    const words = getWords(50);
-    let textBoxes = [];
-    let currWord = current;
-    let renderedWord = 0;
-
-    words.forEach((w) => {
-      // if we're at the current word, highlight it
-      renderedWord === currWord
-        ? textBoxes.push(
-            <Text
-              as="span"
-              m="1"
-              direction="ltr"
-              fontSize="md"
-              className="word-highlighted"
-              color="green.500"
-            >
-              {w + ' '}
-            </Text>
-          )
-        : textBoxes.push(
-            <Text
-              as="span"
-              m="1"
-              direction="ltr"
-              fontSize="md"
-              className="word"
-              color="grey.800"
-            >
-              {w + ' '}
-            </Text>
-          );
-      renderedWord++;
-    });
-
-    return textBoxes;
-  };
+  // const printWords = () => {
+  //   return
+  // };
 
   return (
     <Flex
       className="word-box"
       bg="gray.100"
+      p="1"
       maxW="100%"
       h="auto"
       justify="center"
@@ -72,7 +23,7 @@ const WordBox = () => {
       direction="ltr"
       textAlign={['left']}
     >
-      {highlightCurrWord(0).map((t) => t)}
+      {words.map((t) => t)}
     </Flex>
   );
 };
